Project made with python
In this project were used Selenium, microsoft edge webdriver, and requests-html. Since it is against the youtube policy, i shall advise you not to use the code with the intention to get more viels, your channel can be banned; the intention of this project is to show the cool things one can do with selenium, even in the begginer level.